backend/app/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
pydantic
psycopg2-binary
paho-mqtt
python-dotenv
redis
rq
numpy
scikit-learn
